+++ controller/controlplane/manager_options.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 
-	operatorv2alpha1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v2alpha1"
+	operatorv2beta1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v2beta1"
 
 	"github.com/kong/kong-operator/controller/pkg/log"
 	managercfg "github.com/kong/kong-operator/ingress-controller/pkg/manager/config"
@@ -137,13 +137,13 @@ func WithFeatureGates(logger logr.Logger, featureGates []gwtypes.ControlPlaneFea
 
 // WithReverseSync sets whether configuration is sent to Kong even
 // if the configuration checksum has not changed since previous update.
-func WithReverseSync(state *operatorv2alpha1.ControlPlaneReverseSyncState) managercfg.Opt {
+func WithReverseSync(state *operatorv2beta1.ControlPlaneReverseSyncState) managercfg.Opt {
 	return func(c *managercfg.Config) {
 		if state == nil {
 			return
 		}
 
-		c.EnableReverseSync = *state == operatorv2alpha1.ControlPlaneReverseSyncStateEnabled
+		c.EnableReverseSync = *state == operatorv2beta1.ControlPlaneReverseSyncStateEnabled
 	}
 }
 
@@ -456,7 +456,7 @@ func WithCacheSyncPeriod(period time.Duration) managercfg.Opt {
 }
 
 // WithDataPlaneSyncOptions sets the option to sync Kong configuration with managed dataplanes.
-func WithDataPlaneSyncOptions(syncOptions operatorv2alpha1.ControlPlaneDataPlaneSync) managercfg.Opt {
+func WithDataPlaneSyncOptions(syncOptions operatorv2beta1.ControlPlaneDataPlaneSync) managercfg.Opt {
 	return func(c *managercfg.Config) {
 		if syncOptions.Interval != nil {
 			c.ProxySyncInterval = syncOptions.Interval.Duration
@@ -495,7 +495,7 @@ func WithConfigMapLabelSelector(labelSelector string) managercfg.Opt {
 }
 
 // WithTranslationOptions sets the translation options for the manager.
-func WithTranslationOptions(opts *operatorv2alpha1.ControlPlaneTranslationOptions) managercfg.Opt {
+func WithTranslationOptions(opts *operatorv2beta1.ControlPlaneTranslationOptions) managercfg.Opt {
 	return func(c *managercfg.Config) {
 		if opts == nil {
 			return
@@ -503,15 +503,15 @@ func WithTranslationOptions(opts *operatorv2alpha1.ControlPlaneTranslationOption
 
 		if opts.CombinedServicesFromDifferentHTTPRoutes != nil {
 			c.CombinedServicesFromDifferentHTTPRoutes = (*opts.CombinedServicesFromDifferentHTTPRoutes ==
-				operatorv2alpha1.ControlPlaneCombinedServicesFromDifferentHTTPRoutesStateEnabled)
+				operatorv2beta1.ControlPlaneCombinedServicesFromDifferentHTTPRoutesStateEnabled)
 		}
 		if fc := opts.FallbackConfiguration; fc != nil && fc.UseLastValidConfig != nil {
 			c.UseLastValidConfigForFallback = (*fc.UseLastValidConfig ==
-				operatorv2alpha1.ControlPlaneFallbackConfigurationStateEnabled)
+				operatorv2beta1.ControlPlaneFallbackConfigurationStateEnabled)
 		}
 
 		if opts.DrainSupport != nil {
-			c.EnableDrainSupport = *opts.DrainSupport == operatorv2alpha1.ControlPlaneDrainSupportStateEnabled
+			c.EnableDrainSupport = *opts.DrainSupport == operatorv2beta1.ControlPlaneDrainSupportStateEnabled
 		}
 	}
 }