+++ test/e2e/operator_logs_test.go
@@ -202,7 +202,7 @@ func TestOperatorLogs(t *testing.T) {
 		t.Log("verifying the ControlPlane sub-resource is deleted")
 		if len(controlplanes) != 0 {
 			assert.Eventually(t, func() bool {
-				_, err := clients.OperatorClient.GatewayOperatorV2alpha1().ControlPlanes(testNamespace.Name).Get(ctx, controlplanes[0].Name, metav1.GetOptions{})
+				_, err := clients.OperatorClient.GatewayOperatorV2beta1().ControlPlanes(testNamespace.Name).Get(ctx, controlplanes[0].Name, metav1.GetOptions{})
 				return errors.IsNotFound(err)
 			}, time.Minute, time.Second)
 		}