+++ test/integration/cleaner.go
@@ -9,7 +9,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-// addToCleanup has to be used in tests which work with ControlPlane v2alpha1.
+// addToCleanup has to be used in tests which work with ControlPlane v2beta1.
 // The reason for using this instead of cleaner's methods is that currently,
 // the deletion is being called in
 // https://github.com/Kong/kubernetes-testing-framework/blob/2345c4441ce9ddfef2ded763e9148678b25cd0d1/pkg/clusters/cleanup.go#L214-L217
@@ -18,9 +18,9 @@ import (
 // - https://github.com/kubernetes/kubernetes/issues/80609,
 //
 // and this causes the deletion to somehow delete the object as if it was from
-// the v1beta1 API group which clears v2alpha1 fields
+// the v1beta1 API group which clears v2beta1 fields
 // (because the v1beta1 API group does not have them).
-// Since v2alpha1 ControlPlane have validation rules and objects of this type
+// Since v2beta1 ControlPlane have validation rules and objects of this type
 // get finalizer assigned to clear the in process instance, objects at this
 // point are blocked from being updated or deleted until the CEL validation
 // rules are satisfied.